grpc_docker/requirements.txt
grpcio
grpcio-tools
pyyaml
numpy
networkx